+++ jest.config.js
@@ -1,13 +1,22 @@
 module.exports = {
   testEnvironment: "jsdom",
   coverageDirectory: "./jest-coverage/",
-  coverageReporters: ["json"],
+  coverageReporters: ["lcov"],
   collectCoverage: true,
-  collectCoverageFrom: ["packages/**/src/**/*.(ts|tsx)"],
+  collectCoverageFrom: [
+    "packages/**/src/**/*.(ts|tsx)",
+    "!packages/react/src/*.(ts|tsx)",
+    "!packages/**/src/index.(ts|tsx)",
+    "!packages/table/src/types/*.(ts|tsx)",
+    "!packages/**/src/**/*.d.ts",
+  ],
   moduleFileExtensions: ["js", "jsx", "ts", "tsx"],
   modulePathIgnorePatterns: ["<rootDir>/examples"],
   snapshotSerializers: ["@emotion/jest/serializer"],
-  testPathIgnorePatterns: ["<rootDir>/node_modules/", "<rootDir>/plop-templates/"],
+  testPathIgnorePatterns: [
+    "<rootDir>/node_modules/",
+    "<rootDir>/plop-templates/",
+  ],
   globals: {
     "ts-jest": {
       tsconfig: "tsconfig.json",