+++ packages/cascader/src/cascader.tsx
@@ -1,6 +1,7 @@
 import {
   ForwardedRef,
-  forwardRef, SyntheticEvent,
+  forwardRef,
+  SyntheticEvent,
   useCallback,
   useEffect,
   useRef,
@@ -89,7 +90,7 @@ export const Cascader = forwardRef<HTMLDivElement, CascaderProps<any>>(
     )
 
     const handleVisibleChange = useCallback(
-      (value: any) => {
+      (value: boolean) => {
         if (value !== currentVisible) {
           onVisibleChange?.(value)
           if (!("popupVisible" in props)) {
@@ -176,7 +177,7 @@ export const Cascader = forwardRef<HTMLDivElement, CascaderProps<any>>(
         }
         return {
           text: text || "",
-          disabled: options[options.length - 1]?.disabled,
+          disabled: options?.[options.length - 1]?.disabled,
         }
       },
       [store],