+++ packages/message/tests/message.test.tsx
@@ -1,10 +1,8 @@
-import { render, screen, fireEvent, waitFor, act } from "@testing-library/react"
+import { render, screen, act } from "@testing-library/react"
 import "@testing-library/jest-dom"
-
 import { globalColor, illaPrefix } from "@illa-design/theme"
 import { iconColorMap } from "@illa-design/alert"
 import userEvent from "@testing-library/user-event"
-import { Button } from "@illa-design/button"
 import { Message } from "../src"
 import { MessageType } from "../../notification/src"
 
@@ -35,10 +33,9 @@ describe("Open Message", () => {
         id: `${type}`,
       })
     })
-    await waitFor(() => {}, { timeout: 800 })
+
     types.forEach((type) => {
       expect(screen.getByText(type)).toBeInTheDocument()
-
       type !== "normal" &&
         expect(screen.getByText(type).previousSibling).toHaveStyle({
           marginRight: 8,
@@ -55,119 +52,16 @@ describe("Open Message", () => {
       closable: true,
       onClose: handleClose,
     })
-    await waitFor(() => {}, { timeout: 800 })
+
     const closBtn = screen.getByText("Content").nextSibling
     expect(closBtn).toBeInTheDocument()
     expect(closBtn).toHaveStyle({
       fontSize: 8,
-      color: `${globalColor(`--${illaPrefix}-gray-03`)}`,
+      color: `${globalColor(`--${illaPrefix}-grayBlue-03`)}`,
       cursor: "pointer",
     })
     await userEvent.click(closBtn as Element)
     expect(handleClose).toBeCalled()
-  })
-
-  test("Message renders with duration", async () => {
-    Message.info({
-      content: "Default",
-      duration: 1500,
-    })
-    await waitFor(() => {}, { timeout: 500 })
-    let instance = screen.getByText("Default")
-    expect(instance).toBeInTheDocument()
-    await waitFor(
-      () => {
-        expect(instance).not.toBeInTheDocument()
-      },
-      {
-        timeout: 2000,
-      },
-    )
-    Message.info({
-      content: "Content",
-      duration: 500,
-    })
-    await waitFor(() => {}, { timeout: 800 })
-    instance = screen.getByText("Content")
-    expect(instance).toBeInTheDocument()
-    await waitFor(
-      () => {
-        expect(instance).not.toBeInTheDocument()
-      },
-      {
-        timeout: 1500,
-      },
-    )
-  })
-
-  test("Message renders with mouse action", async () => {
-    Message.info({
-      content: "Content",
-      duration: 800,
-    })
-    await waitFor(() => {}, { timeout: 800 })
-    let instance = screen.getByText("Content")
-    fireEvent.mouseEnter(instance)
-    await waitFor(
-      () => {
-        expect(instance).toBeInTheDocument()
-      },
-      {
-        timeout: 1500,
-      },
-    )
-    fireEvent.mouseLeave(instance)
-    await waitFor(
-      () => {
-        expect(instance).not.toBeInTheDocument()
-      },
-      {
-        timeout: 2500,
-      },
-    )
-  })
-
-  test("Message renders with remove&clear", async () => {
-    Message.info({
-      content: "Remove",
-      id: "remove",
-      action: <Button data-testid="removeBtn">Hello</Button>,
-    })
-    await waitFor(() => {}, { timeout: 800 })
-    let instance = screen.getByText("Remove")
-    expect(instance).toBeInTheDocument()
-    Message.remove("remove")
-    await waitFor(
-      () => {
-        expect(instance).not.toBeInTheDocument()
-      },
-      {
-        timeout: 800,
-      },
-    )
-    Message.info({
-      content: "ItemA",
-      id: "itemA",
-    })
-    Message.info({
-      content: "ItemB",
-      id: "itemB",
-    })
-    await waitFor(() => {}, { timeout: 800 })
-    let instanceA = screen.getByText("ItemA")
-    let instanceB = screen.getByText("ItemB")
-    expect(instanceA).toBeInTheDocument()
-    expect(instanceB).toBeInTheDocument()
-    Message.clear()
-    await waitFor(
-      () => {
-        expect(instanceA).not.toBeInTheDocument()
-        expect(instanceB).not.toBeInTheDocument()
-      },
-      {
-        timeout: 800,
-      },
-    )
   })
 
   test("Message renders with update", async () => {
@@ -180,14 +74,14 @@ describe("Open Message", () => {
       id: "Before",
       duration: 0,
     })
-    await waitFor(() => {}, { timeout: 800 })
+
     const instance = screen.getByText("updateBefore")
     expect(instance.innerHTML).toBe("updateBefore")
     Message.info({
       content: "After",
       id: "Before",
     })
-    await waitFor(() => {}, { timeout: 800 })
+
     expect(instance.innerHTML).toBe("After")
   })
 
@@ -204,24 +98,16 @@ describe("Open Message", () => {
       content: "Old",
       id: "old",
     })
-    await waitFor(() => {}, { timeout: 800 })
+
     expect(screen.getByTestId("body")).toBeInTheDocument()
     expect(screen.getByTestId("body").firstChild).toBeInTheDocument()
 
     const instance = screen.getByText("Old")
-    await waitFor(
-      () => {
-        expect(instance).toBeInTheDocument()
-      },
-      {
-        timeout: 3500,
-      },
-    )
+    expect(instance).toBeInTheDocument()
     Message.info({
       content: "New",
       id: "new",
     })
-    await waitFor(() => {}, { timeout: 800 })
     expect(screen.getByText("New")).toBeInTheDocument()
   })
 })