+++ packages/switch/tests/switch.test.tsx
@@ -59,7 +59,7 @@ test("Switch renders with different colorScheme in different status", () => {
 
   render(<Switch data-testid="test-unchecked" colorScheme={"black"} />)
   expect(screen.getByTestId("test-unchecked")).toHaveStyle({
-    "background-color": "#d6d6d6",
+    "background-color": globalColor(`--${illaPrefix}-grayBlue-06`),
   })
 })
 
@@ -74,7 +74,7 @@ test("Switch renders with status changed", () => {
     />,
   )
   expect(screen.getByTestId("test-status-changed")).toHaveStyle({
-    "background-color": "#d6d6d6",
+    "background-color": globalColor(`--${illaPrefix}-grayBlue-06`),
   })
   const target = screen.getByTestId("test-status-changed")
   fireEvent.click(target)
@@ -135,7 +135,7 @@ test("Switch renders with disabled", () => {
   )
   expect(screen.getByTestId("test-disabled")).toHaveStyle({
     cursor: "not-allowed",
-    "background-color": "#ebebeb",
+    "background-color": globalColor(`--${illaPrefix}-grayBlue-08`),
   })
   fireEvent.click(screen.getByTestId("test-disabled"))
   expect(onChangeEvent).not.toBeCalled()