+++ packages/tooltip/tests/tolltip.e2e.tsx
@@ -1,0 +1,10 @@
+import { mount, unmount } from "@cypress/react"
+import "@testing-library/cypress"
+import { Tooltip } from "../src"
+
+it("Tooltip render correctly", () => {
+  mount(<Tooltip content="Tooltip">Test</Tooltip>)
+  cy.findByText("Test").trigger("mouseover")
+  cy.findByText("Tooltip").should("exist")
+  unmount()
+})