+++ scripts/merge-coverage.sh
@@ -1,5 +1,11 @@
 #!/usr/bin/env bash
 mkdir reports || true
-cp cypress-coverage/coverage-final.json reports/from-cypress.json && cp jest-coverage/coverage-final.json reports/from-jest.json
-npx nyc merge reports && mv coverage.json .nyc_output/out.json
-npx nyc report --reporter lcov --reporter text --report-dir coverage
+cp cypress-coverage/lcov.info reports/from-cypress.info && cp jest-coverage/lcov.info reports/from-jest.info
+mkdir coverage || true
+
+cypressInfoPath=./reports/from-cypress.info
+jestInfoPath=./reports/from-jest.info
+
+./scripts/mergeLcov.perl -a $cypressInfoPath -a $jestInfoPath -o ./coverage/lcov.info
+
+./scripts/genHtml.perl --legend --branch-coverage -t merged-coverage-overview $cypressInfoPath $jestInfoPath --o=./coverage/lcov-report -c=./scripts/lcov-style.css