+++ packages/date-picker/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @illa-design/date-picker
 
+## 0.0.3
+
+### Patch Changes
+
+- fix some bugs
+
+- Updated dependencies []:
+  - @illa-design/calendar@0.0.3
+  - @illa-design/icon@0.0.3
+  - @illa-design/select@0.0.3
+  - @illa-design/system@0.0.3
+  - @illa-design/theme@0.0.3
+  - @illa-design/time-picker@0.0.3
+  - @illa-design/trigger@0.0.3
+
+## 0.0.2
+
+### Patch Changes
+
+- update all components version
+
+- Updated dependencies []:
+  - @illa-design/calendar@0.0.2
+  - @illa-design/icon@0.0.2
+  - @illa-design/select@0.0.2
+  - @illa-design/system@0.0.2
+  - @illa-design/theme@0.0.2
+  - @illa-design/time-picker@0.0.2
+  - @illa-design/trigger@0.0.2
+
 ## 0.0.1
 
 ### Patch Changes