+++ packages/modal/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @illa-design/modal
 
+## 0.0.3
+
+### Patch Changes
+
+- fix some bugs
+
+- Updated dependencies []:
+  - @illa-design/alert@0.0.3
+  - @illa-design/button@0.0.3
+  - @illa-design/config-provider@0.0.3
+  - @illa-design/icon@0.0.3
+  - @illa-design/system@0.0.3
+  - @illa-design/theme@0.0.3
+
+## 0.0.2
+
+### Patch Changes
+
+- update all components version
+
+- Updated dependencies []:
+  - @illa-design/alert@0.0.2
+  - @illa-design/button@0.0.2
+  - @illa-design/config-provider@0.0.2
+  - @illa-design/icon@0.0.2
+  - @illa-design/system@0.0.2
+  - @illa-design/theme@0.0.2
+
 ## 0.0.1
 
 ### Patch Changes