+++ packages/react/CHANGELOG.md
@@ -1,5 +1,139 @@
 # @illa-design/react
 
+## 0.0.3
+
+### Patch Changes
+
+- fix some bugs
+
+- Updated dependencies []:
+  - @illa-design/affix@0.0.3
+  - @illa-design/alert@0.0.3
+  - @illa-design/anchor@0.0.3
+  - @illa-design/avatar@0.0.3
+  - @illa-design/back-top@0.0.3
+  - @illa-design/badge@0.0.3
+  - @illa-design/button@0.0.3
+  - @illa-design/calendar@0.0.3
+  - @illa-design/card@0.0.3
+  - @illa-design/checkbox@0.0.3
+  - @illa-design/collapse@0.0.3
+  - @illa-design/comment@0.0.3
+  - @illa-design/config-provider@0.0.3
+  - @illa-design/date-picker@0.0.3
+  - @illa-design/description@0.0.3
+  - @illa-design/divider@0.0.3
+  - @illa-design/drawer@0.0.3
+  - @illa-design/empty@0.0.3
+  - @illa-design/grid@0.0.3
+  - @illa-design/icon@0.0.3
+  - @illa-design/image@0.0.3
+  - @illa-design/input@0.0.3
+  - @illa-design/input-number@0.0.3
+  - @illa-design/input-tag@0.0.3
+  - @illa-design/link@0.0.3
+  - @illa-design/list@0.0.3
+  - @illa-design/menu@0.0.3
+  - @illa-design/message@0.0.3
+  - @illa-design/modal@0.0.3
+  - @illa-design/notification@0.0.3
+  - @illa-design/page-header@0.0.3
+  - @illa-design/pagination@0.0.3
+  - @illa-design/popconfirm@0.0.3
+  - @illa-design/popover@0.0.3
+  - @illa-design/progress@0.0.3
+  - @illa-design/radio@0.0.3
+  - @illa-design/rate@0.0.3
+  - @illa-design/result@0.0.3
+  - @illa-design/select@0.0.3
+  - @illa-design/skeleton@0.0.3
+  - @illa-design/slider@0.0.3
+  - @illa-design/space@0.0.3
+  - @illa-design/spin@0.0.3
+  - @illa-design/statistic@0.0.3
+  - @illa-design/steps@0.0.3
+  - @illa-design/switch@0.0.3
+  - @illa-design/system@0.0.3
+  - @illa-design/table@0.0.3
+  - @illa-design/tabs@0.0.3
+  - @illa-design/tag@0.0.3
+  - @illa-design/theme@0.0.3
+  - @illa-design/time-picker@0.0.3
+  - @illa-design/timeline@0.0.3
+  - @illa-design/tooltip@0.0.3
+  - @illa-design/tree@0.0.3
+  - @illa-design/tree-common@0.0.3
+  - @illa-design/tree-select@0.0.3
+  - @illa-design/trigger@0.0.3
+  - @illa-design/typography@0.0.3
+  - @illa-design/upload@0.0.3
+
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @illa-design/affix@0.0.2
+  - @illa-design/alert@0.0.2
+  - @illa-design/anchor@0.0.2
+  - @illa-design/avatar@0.0.2
+  - @illa-design/back-top@0.0.2
+  - @illa-design/badge@0.0.2
+  - @illa-design/button@0.0.2
+  - @illa-design/calendar@0.0.2
+  - @illa-design/card@0.0.2
+  - @illa-design/checkbox@0.0.2
+  - @illa-design/collapse@0.0.2
+  - @illa-design/comment@0.0.2
+  - @illa-design/config-provider@0.0.2
+  - @illa-design/date-picker@0.0.2
+  - @illa-design/description@0.0.2
+  - @illa-design/divider@0.0.2
+  - @illa-design/drawer@0.0.2
+  - @illa-design/empty@0.0.2
+  - @illa-design/grid@0.0.2
+  - @illa-design/icon@0.0.2
+  - @illa-design/image@0.0.2
+  - @illa-design/input@0.0.2
+  - @illa-design/input-number@0.0.2
+  - @illa-design/input-tag@0.0.2
+  - @illa-design/link@0.0.2
+  - @illa-design/list@0.0.2
+  - @illa-design/menu@0.0.2
+  - @illa-design/message@0.0.2
+  - @illa-design/modal@0.0.2
+  - @illa-design/notification@0.0.2
+  - @illa-design/page-header@0.0.2
+  - @illa-design/pagination@0.0.2
+  - @illa-design/popconfirm@0.0.2
+  - @illa-design/popover@0.0.2
+  - @illa-design/progress@0.0.2
+  - @illa-design/radio@0.0.2
+  - @illa-design/rate@0.0.2
+  - @illa-design/result@0.0.2
+  - @illa-design/select@0.0.2
+  - @illa-design/skeleton@0.0.2
+  - @illa-design/slider@0.0.2
+  - @illa-design/space@0.0.2
+  - @illa-design/spin@0.0.2
+  - @illa-design/statistic@0.0.2
+  - @illa-design/steps@0.0.2
+  - @illa-design/switch@0.0.2
+  - @illa-design/system@0.0.2
+  - @illa-design/table@0.0.2
+  - @illa-design/tabs@0.0.2
+  - @illa-design/tag@0.0.2
+  - @illa-design/theme@0.0.2
+  - @illa-design/time-picker@0.0.2
+  - @illa-design/timeline@0.0.2
+  - @illa-design/tooltip@0.0.2
+  - @illa-design/tree@0.0.2
+  - @illa-design/tree-select@0.0.2
+  - @illa-design/trigger@0.0.2
+  - @illa-design/typography@0.0.2
+  - @illa-design/upload@0.0.2
+  - @illa-design/tree-common@0.0.2
+
 ## 0.0.1
 
 ### Patch Changes