+++ packages/slider/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @illa-design/slider
 
+## 0.0.3
+
+### Patch Changes
+
+- fix some bugs
+
+- Updated dependencies []:
+  - @illa-design/input-number@0.0.3
+  - @illa-design/system@0.0.3
+  - @illa-design/theme@0.0.3
+  - @illa-design/trigger@0.0.3
+
+## 0.0.2
+
+### Patch Changes
+
+- update all components version
+
+- Updated dependencies []:
+  - @illa-design/input-number@0.0.2
+  - @illa-design/system@0.0.2
+  - @illa-design/theme@0.0.2
+  - @illa-design/trigger@0.0.2
+
 ## 0.0.1
 
 ### Patch Changes