+++ packages/theme/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @illa-design/theme
 
+## 0.0.3
+
+### Patch Changes
+
+- fix some bugs
+
+- Updated dependencies []:
+  - @illa-design/system@0.0.3
+
+## 0.0.2
+
+### Patch Changes
+
+- update all components version
+
+- Updated dependencies []:
+  - @illa-design/system@0.0.2
+
 ## 0.0.1
 
 ### Patch Changes