+++ packages/tree-select/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @illa-design/treeselect
 
+## 0.0.3
+
+### Patch Changes
+
+- fix some bugs
+
+- Updated dependencies []:
+  - @illa-design/empty@0.0.3
+  - @illa-design/select@0.0.3
+  - @illa-design/system@0.0.3
+  - @illa-design/theme@0.0.3
+  - @illa-design/tree@0.0.3
+  - @illa-design/tree-common@0.0.3
+
+## 0.0.2
+
+### Patch Changes
+
+- update all components version
+
+- Updated dependencies []:
+  - @illa-design/empty@0.0.2
+  - @illa-design/select@0.0.2
+  - @illa-design/system@0.0.2
+  - @illa-design/theme@0.0.2
+  - @illa-design/tree@0.0.2
+  - @illa-design/tree-common@0.0.2
+
 ## 0.0.1
 
 ### Patch Changes