+++ packages/upload/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @illa-design/avatar
 
+## 0.0.3
+
+### Patch Changes
+
+- fix some bugs
+
+- Updated dependencies []:
+  - @illa-design/config-provider@0.0.3
+  - @illa-design/icon@0.0.3
+  - @illa-design/image@0.0.3
+  - @illa-design/list@0.0.3
+  - @illa-design/progress@0.0.3
+  - @illa-design/system@0.0.3
+  - @illa-design/theme@0.0.3
+
+## 0.0.2
+
+### Patch Changes
+
+- update all components version
+
+- Updated dependencies []:
+  - @illa-design/config-provider@0.0.2
+  - @illa-design/icon@0.0.2
+  - @illa-design/image@0.0.2
+  - @illa-design/list@0.0.2
+  - @illa-design/progress@0.0.2
+  - @illa-design/system@0.0.2
+  - @illa-design/theme@0.0.2
+
 ## 0.0.1
 
 ### Patch Changes